requirements.txt
selenium
psycopg2-binary
python-dotenv
schedule
beautifulsoup4
requests